coffee chat screen updates
