+++ src/screens/CoffeeChat/CoffeeChat.css
@@ -1,22 +1,12 @@
  @import url('https://fonts.googleapis.com/css2?family=Itim&display=swap');
 
-/* * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: Helvetica, 'Itim';
-} */
-
-/* body {
-    background-color: #EDECE4 !important;
-    color: #000000;
-    padding: 20px;
-} */
-
 .container {
     max-width: 1200px;
     width: 100%;
     margin: 0 auto;
+    max-height:1200px;
+    height: 100%;
+    overflow-y: auto;
 }
 
 header {
@@ -63,7 +53,7 @@ h1 {
     gap: 15px;
     margin-bottom: 20px;
     overflow-x: auto;
-    padding-bottom: 10px;
+    padding-bottom: 40px;
     max-width: 800px;
     width: 100%;
 }
@@ -72,12 +62,13 @@ h1 {
     background-color: #D0E1D1;
     border: 1px solid #10300E;
     color: #10300E;
-    padding: 8px 16px;
+    padding: 10px 16px;
     border-radius: 20px;
     font-size: 14px;
     font-weight: 600;
     cursor: pointer;
     white-space: nowrap;
+    text-align: center;
 }
 
 .category.active {
@@ -90,6 +81,8 @@ h1 {
     grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
     gap: 20px;
     margin-bottom: 30px;
+    max-width: 800px;
+    width: 100%;
 }
 
 .student-card {
@@ -100,6 +93,7 @@ h1 {
     transition: transform 0.3s ease;
     display: flex;
     flex-direction: column;
+    align-items: center;
 }
 
 .student-card:hover {
@@ -109,14 +103,13 @@ h1 {
 .student-topbar {
     background-color: #3B6838;
     color: #fff;
-    padding: 15px;
+    padding: 10px;
     width: 100%;
     flex-shrink: 0;
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    text-align: center;
 }
 
 .student-name {
